Support multi inputs for ifelse fallback
