+++ tests/test_executor/test_ifelse_fallback.py
@@ -20,5 +20,21 @@ class TestIfElse(TestCaseBase):
         self.assert_results(ifelse_func, x, y)
 
 
+def multi_output(x: paddle.Tensor):
+    m = x + 1
+    if x > 0:
+        return m
+    else:
+        return 2 * m
+
+
+class TestExecutor(TestCaseBase):
+    def test_simple(self):
+        x = paddle.to_tensor(2)
+        self.assert_results(multi_output, x)
+        x = paddle.to_tensor(-2)
+        self.assert_results(multi_output, x)
+
+
 if __name__ == "__main__":
     unittest.main()
